Initial Commit
Wordle Prune is a tool that allows users to "cheat" in wordle.

In extents of cheating it is not the best as there are ways to
definitively find out the correct word as it is exposed in the
source code.

This was done as a fun project that is only half-cheating.